+++ Source/ShinyMathRocksDefOf.cs
@@ -10,6 +10,7 @@ namespace ShinyMathRocks
         public static JobDef SMR_RollDice;
         public static ThoughtDef SMR_DiceRollMood;
         public static SoundDef SMR_DiceClatter;
+        public static HediffDef SMR_DiceGoblinStatus;
 
         static ShinyMathRocksDefOf()
         {